ULTIMA4: Move timed event code into it's own file
